requirements.txt
fastapi
uvicorn
pandas
SQLalchemy
pwdlib
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
pwdlib[argon2]
redis
aiosmtplib
pydantic-settings
slowapi
joblib
asyncio
scikit-learn==1.6.1
xgboost